update: refactor user related services and models
