+++ Backend/Model/User/UserGoals.cs
@@ -2,18 +2,13 @@ using System.Text.Json.Serialization;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 
-namespace AiRise.Models
+namespace AiRise.Models.User
 {
-    public class UserHealthData
+    public class UserGoals
     {
         [BsonId]
         [BsonRepresentation(BsonType.ObjectId)]
         public string? Id { get; set; }
-
-        // Populate with json array from client side app
-        [BsonElement("health_data_id")]
-        [JsonPropertyName("health_data_id")]
-        public string HealthDataId { get; set; } = string.Empty;
 
         [BsonElement("calories")]
         [JsonPropertyName("calories")]
@@ -23,8 +18,8 @@ namespace AiRise.Models
         [JsonPropertyName("steps")]
         public int Steps { get; set; } = 0;
 
-        [BsonElement("sleep")]
-        [JsonPropertyName("sleep")]
-        public int Sleep { get; set; } = 0;
+        [BsonElement("weight")]
+        [JsonPropertyName("weight")]
+        public int Weight { get; set; } = 0;
     }
 }
